requirements.txt
requests
beautifulsoup4
numpy
faiss-cpu
ollama
fastapi 
uvicorn
jinja2
